% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{Generate report via UI}
\usage{
generateReport(template = NULL, parameters = reviewParams(), ...)
}
\arguments{
\item{template}{An alternative template, if required}

\item{parameters}{Parameters to be provided to the report, based on \code{\link[=reviewParams]{reviewParams()}}}

\item{...}{Additional arguments to \code{\link[rmarkdown:render]{rmarkdown::render()}}}
}
\description{
Generate report via UI
}
\examples{
generateReport()
}
