% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.UI.R
\name{generateReport.UI}
\alias{generateReport.UI}
\title{Generate report via UI}
\usage{
generateReport.UI(template = NULL, ...)
}
\arguments{
\item{template}{An alternative template, if required}

\item{...}{Additional arguments to \code{\link[rmarkdown:render]{rmarkdown::render()}}}
}
\description{
Generate report via UI
}
\examples{
\dontrun{
generateReport.UI()
}
}
