% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewParams.R
\name{reviewParams}
\alias{reviewParams}
\title{Specify review parameters}
\usage{
reviewParams(package = "PackageReviewR", reviewAuthor = "Steph Locke",
  reviewDate = Sys.Date(), packageDeps = TRUE, functionDeps = TRUE,
  codeCoverage = TRUE)
}
\arguments{
\item{package}{Package to be reviews}

\item{reviewAuthor}{Name of reviewer}

\item{reviewDate}{Date of review}

\item{packageDeps}{Include package dependencies information}

\item{functionDeps}{Include function dependency information}

\item{codeCoverage}{Include code coverage information}
}
\value{
A list of parameters to be used in \code{\link[=generateReport]{generateReport()}}
}
\description{
Specify review parameters
}
\examples{
reviewParams()
reviewParams(reviewAuthor="New author")
}
